remove the UA_NodeStore_release function; zero-copy reads
its inteded use case is for the multithreaded server. but with the current RCU mechanism, not even that is needed.
we just enter a quiescent section at the end of each processed job.

this enables zero-copy reads. since we don't have to "release" lots of nodes after the binary encoding (where we currently don't know what these nodes would have been).